requirements.txt
google-adk==1.14.0
google-cloud-logging==3.11.4
google-generativeai
python-dotenv==1.0.1